converted what was client-side api routes to NextJS's serverless routes.  Also reorganised url path of the search page to now be '/search/[...list]' as the previous structure caused a bug where the dynamic routes would pick up the serverless api folder and think anything accessing the route was refering to a page instead of a route.  Not good. Works perfectly now, though.  Yay.
