+++ package.json
@@ -16,6 +16,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.6.7",
+    "swr": "^0.5.6",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
